pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "amplifier-hive-slack"
version = "0.1.0"
description = "Multi-agent collaboration via Slack, powered by Amplifier"
requires-python = ">=3.10"
dependencies = [
    "amplifier-foundation @ git+https://github.com/microsoft/amplifier-foundation@main",
    "slack-bolt>=1.18",
    "aiohttp>=3.9",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[project.scripts]
hive-slack = "hive_slack.main:cli"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/hive_slack"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
